:art: Refresh associated blocks after find-replacing and rolling back doc https://github.com/siyuan-note/siyuan/issues/12439 https://github.com/siyuan-note/siyuan/issues/12438
