+++ kernel/task/queue.go
@@ -134,6 +134,7 @@ const (
 	AssetContentDatabaseIndexCommit = "task.asset.database.index.commit"   // 资源文件数据库索引提交
 	CacheVirtualBlockRef            = "task.cache.virtualBlockRef"         // 缓存虚拟块引用
 	ReloadAttributeView             = "task.reload.attributeView"          // 重新加载属性视图
+	ReloadProtyle                   = "task.reload.protyle"                // 重新加载编辑器
 	SetRefDynamicText               = "task.ref.setDynamicText"            // 设置引用的动态锚文本
 	SetDefRefCount                  = "task.def.setRefCount"               // 设置定义的引用计数
 	PushMsg                         = "task.push.msg"                      // 推送消息
@@ -151,6 +152,7 @@ var uniqueActions = []string{
 	AssetContentDatabaseIndexFull,
 	AssetContentDatabaseIndexCommit,
 	ReloadAttributeView,
+	ReloadProtyle,
 	SetRefDynamicText,
 	SetDefRefCount,
 }
